\name{geneCounts}
\Rdversion{1}
\alias{geneCounts}
\title{Gene counts.}
\description{
Get gene counts from GRanges object
}
\usage{
geneCounts(Anno)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{Anno}{An annotation object of type \code{GRanges} which has been simplified and had a counts matrix merged.}
}
\value{
An object of type \code{matrix}
}
\author{ Ellis Patrick }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
