\name{mergeCounts}
\Rdversion{1}
\alias{mergeCounts}
\title{Merge Counts}
\description{
Merges a counts matrix with a GRanges object.
}
\usage{
mergeCounts(c, GRanges)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{c}{a matrix of counts}
  \item{cut}{An object of type \code{GRanges} }
}
  \details{

Merges a GRanges object with a counts matrix.
}
\value{
An object of type \code{GRanges}
}
\author{ Ellis Patrick }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
