\name{pMimHTML}
\Rdversion{1}
\alias{pMimHTML}
\title{Output HTML files for pMim}
\description{
This function outputs HTML files for the pMim function
}
\usage{
pMimHTML(output,dir = 'mirpathways',cutoff = 0.05,filename = 'pMimResults.html',GeneSymbol = NULL,outputHTML=FALSE)}
\arguments{
\item{output}{A list of output that was generated by pMim.}
\item{dir}{The directory you would like the information about the genes in each mirpathway to be written.}
  \item{cut-off}{Mir-pathways must have an intergrative score greater than this to be included in results.}
\item{filename}{The name you would like you html file to be called.}
\item{GeneSymbol}{A named vector that can transform the rownames of the Gene Data matrix to Gene Symbols.}
\item{outputHTML}{If TRUE this outputs a call to gvisTable.}
}
  \details{
This function write the results from a call to pMim to html files.
}
\value{
If outputHTML is TRUE this outputs the results from pMim as a call to gvisTable. Otherwise this simple writes the results to a html file.
}
\author{ Ellis Patrick }
\examples{}

